get rid of /blog
