+++ pages/index.jsx
@@ -1,9 +1,20 @@
-import DefaultLayout from '../_layouts/default'
+import { getConfig, getAllPosts } from '../api'
+import Layout from '../_layouts/blog/landingPage'
 
-const Home = () => (
-  <DefaultLayout title="Home">
-    This is home
-  </DefaultLayout>
-)
+export default function Blog({ description, posts }) {
+  return (
+    <Layout title="Blog" description={description} {...{ posts }} />
+  )
+}
 
-export default Home
+export async function getStaticProps() {
+  const config = await getConfig()
+  const allPosts = await getAllPosts()
+  return {
+    props: {
+      posts: allPosts,
+      title: config.title,
+      description: config.description
+    }
+  }
+}
